requirements.txt
Pillow
numpy
scipy
svgwrite
pywebview